+++ budget_app/model/pages/ExpensePage.java
@@ -55,9 +55,13 @@ public class ExpensePage extends Page {
     private void deleteExpense() {
         DatabaseConnector databaseConnector = new DatabaseConnector();
         int expenseID;
+        System.out.println("To cancel, just type \"0\".");
         System.out.println("Which Expense would you like to delete? Enter the corresponding expense id below.");
         try {
             expenseID = takeUserInputInt();
+            if(expenseID == 0){
+                return;
+            }
             String sql = "DELETE FROM expenses WHERE expense_id = " + expenseID + ";";
             databaseConnector.openConnection();
             databaseConnector.executeDeleteStatement(sql);
@@ -71,9 +75,13 @@ public class ExpensePage extends Page {
     private void updateExpense() {
         DatabaseConnector databaseConnector = new DatabaseConnector();
         int expenseID;
+        System.out.println("To cancel, just type \"0\".");
         System.out.println("Which Expense would you like to update? Enter the corresponding expense id below: ");
         try {
             expenseID = takeUserInputInt();
+            if (expenseID == 0){
+                return;
+            }
             databaseConnector.openConnection();
             databaseConnector.resultSet = databaseConnector.executeQuery("SELECT expenses.* FROM expenses" +
                     " INNER JOIN users" +
